beacon fixes
