+++ forest/shared/Cargo.toml
@@ -15,6 +15,7 @@ byte-unit = "4.0"
 chrono.workspace = true
 console-subscriber.workspace = true
 directories.workspace = true
+forest_beacon.workspace = true
 forest_chain_sync.workspace = true
 forest_db.workspace = true
 forest_libp2p.workspace = true
