+++ networks/src/drand.rs
@@ -2,10 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use std::borrow::Cow;
+use std::time::Duration;
 
 use forest_beacon::ChainInfo;
 use forest_beacon::DrandConfig;
 use forest_beacon::DrandNetwork;
+use forest_beacon::TokioConfig;
 
 pub(super) static DRAND_MAINNET: DrandConfig<'static> = DrandConfig {
     server: "https://api.drand.sh",
@@ -18,6 +20,7 @@ pub(super) static DRAND_MAINNET: DrandConfig<'static> = DrandConfig {
         group_hash: Cow::Borrowed("176f93498eac9ca337150b46d21dd58673ea4e3581185f869672e59fa4cb390a"),
     },
     network_type: DrandNetwork::Mainnet,
+    tokio: TokioConfig { worker_threads_number: 4, max_number_of_blocking_threads: 512, blocking_thread_keep_alive_timeout: Duration::from_secs(10), thread_stack_size: 2 * 1024 * 1024, global_queue_interval: 61 },
 };
 
 pub(super) static DRAND_INCENTINET: DrandConfig<'static> = DrandConfig {
@@ -31,4 +34,5 @@ pub(super) static DRAND_INCENTINET: DrandConfig<'static> = DrandConfig {
         group_hash: Cow::Borrowed("d9406aaed487f7af71851b4399448e311f2328923d454e971536c05398ce2d9b"),
     },
     network_type: DrandNetwork::Incentinet,
+    tokio: TokioConfig { worker_threads_number: 4, max_number_of_blocking_threads: 512, blocking_thread_keep_alive_timeout: Duration::from_secs(10), thread_stack_size: 2 * 1024 * 1024, global_queue_interval: 61 },
 };
